requirements.txt
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
aiohttp>=3.8.0
numpy>=1.21.0
tqdm>=4.64.0
scikit-learn>=1.0.0
networkx>=2.6.0
scipy>=1.10.0
sentence-transformers>=5.1.0
streamlit
pymongo
langchain
langchain-openai
python-dotenv
faiss-cpu
langchain-upstage
openai
langgraph

